move membership types to new file
