+++ frontend-react/src/pages/submissions/SubmissionTable.test.tsx
@@ -13,7 +13,7 @@ import { mockSessionContext } from "../../contexts/__mocks__/SessionContext";
 import {
     MembershipController,
     MemberType,
-} from "../../hooks/UseOktaMemberships";
+} from "../../hooks/types/UseOktaMembershipsTypes";
 import { SessionController } from "../../hooks/UseSessionStorage";
 
 import SubmissionTable from "./SubmissionTable";
